feat(combobox): fix minor issues
